Safari-609.1.11.1/LayoutTests/crypto/subtle/ecdsa-generate-key-sign-verify-p521-expected.txt

Test ECDSA signing/verifying operations with generated P-521 keys

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating keys
Signing
PASS signature.byteLength is 132
Verifying
PASS verified is true
PASS successfullyParsed is true

TEST COMPLETE
